renamed component "Content" to "Main"
