+++ src/App.css
@@ -6,7 +6,7 @@
   grid-auto-rows: minmax(50px, auto);
 
   font-family: "Helvetica Neue";
-  //font-weight: 150;
+  /* font-weight: 150; */
 }
 
 h1 {
